requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.5.0
psycopg2-binary==2.9.7
redis==4.6.0
web3==6.11.0
eth-account==0.9.0
cryptography==41.0.7
# boto3==1.28.57  # Uncomment for AWS S3 or Digital Ocean Spaces
opencv-python==4.8.1.78
Pillow==10.0.1
requests==2.31.0
google-auth==2.23.4
googlemaps==4.10.0
stripe==7.0.0
paypalrestsdk==1.13.2
twilio==8.10.0
qrcode==7.4.2
python-dotenv==1.0.0
numpy==1.25.2
anthropic==0.7.8
geopy==2.3.0
reverse-geocode==1.4.0
marshmallow==3.20.1
python-magic==0.4.27
bleach==6.1.0
